I'm back, currently working on spotify api
