+++ src/App.js
@@ -1,7 +1,7 @@
 import Home from './Webpages/Home';
-import Databases from './Webpages/Database_Usage';
-import YouTubeAPI from './Webpages/YouTube_API';
-import SpotifyAPI from './Webpages/Spotify_API';
+import Databases from './Webpages/DatabaseUsage';
+import YouTubeAPI from './Webpages/YouTubeAPI';
+import SpotifyAPI from './Webpages/SpotifyAPI';
 import Contact from './Webpages/Contact';
 import PageNotFound from './Webpages/404Page'
 import SpotifySongs from "./Webpages/SpotifySongs"
