+++ src/Webpages/DatabaseUsage.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Databases(){
+    return(
+        <div>
+            <h1>Databases</h1>
+        </div>
+    )
+}
+
+export default Databases;
