--- src/Webpages/Database_Usage.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function Databases(){
-    return(
-        <div>
-            <h1>Databases</h1>
-        </div>
-    )
-}
-
-export default Databases;
