--- src/Webpages/YouTube_API.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function YouTubeAPI(){
-    return(
-        <div>
-            <h1>YouTube</h1>
-        </div>
-    )
-}
-
-export default YouTubeAPI;
